Create a directory per logical group of e2e test
