+++ e2e/simple/BUILD
@@ -1,0 +1,17 @@
+# gazelle:exclude simple_test.go
+
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("@com_github_bazelbuild_bazel_integration_testing//go:bazel_integration_test.bzl", "bazel_go_integration_test")
+
+bazel_go_integration_test(
+    name = "go_default_test",
+    srcs = ["simple_test.go"],
+    data = [
+        "//ibazel",
+    ],
+    importpath = "github.com/bazelbuild/bazel-watcher/e2e/simple",
+    deps = [
+        "//e2e:go_default_library",
+        "@com_github_bazelbuild_bazel_integration_testing//go:go_default_library",
+    ],
+)
